Renamed index to indexExchange,for legacy compatibility.
