+++ adapters/indexExchange/index.go
@@ -1,4 +1,4 @@
-package index
+package indexExchange
 
 import (
 	"bytes"
