+++ adapters/indexExchange/index_test.go
@@ -1,4 +1,4 @@
-package index
+package indexExchange
 
 import (
 	"context"
