+++ adapters/indexExchange/info.yaml
