+++ pbs_light_test.go
@@ -482,6 +482,7 @@ func TestWriteAuctionError(t *testing.T) {
 }
 
 func ensureHasKey(t *testing.T, data map[string]json.RawMessage, key string) {
+	t.Helper()
 	if _, ok := data[key]; !ok {
 		t.Errorf("Expected map to produce a schema for adapter: %s", key)
 	}
